fix: format article markdown in Event component (Article Event)
Changes:
- extracted article formatting code from Article Page into FormatService
- using FormatService in Article Page to render the full article
- using FormatService in ArticleEvent component to render an article/preview in profile/feeds etc
